requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.0.0
urllib3>=2.0.0
charset-normalizer>=3.0.0 